refactor(ui): split configuration adapter ownership
